Header view refactor.
